Breakup test files
